+++ angular-workspace/projects/ni/nimble-angular/label-provider/table/nimble-label-provider-table.directive.ts
@@ -18,8 +18,6 @@ export class NimbleLabelProviderTableDirective {
         return this.elementRef.nativeElement.groupCollapse;
     }
 
-    // Renaming because property should have camel casing, but attribute should not
-    // eslint-disable-next-line @angular-eslint/no-input-rename
     @Input('group-collapse') public set groupCollapse(value: string | undefined) {
         this.renderer.setProperty(this.elementRef.nativeElement, 'groupCollapse', value);
     }
@@ -28,8 +26,6 @@ export class NimbleLabelProviderTableDirective {
         return this.elementRef.nativeElement.groupExpand;
     }
 
-    // Renaming because property should have camel casing, but attribute should not
-    // eslint-disable-next-line @angular-eslint/no-input-rename
     @Input('group-expand') public set groupExpand(value: string | undefined) {
         this.renderer.setProperty(this.elementRef.nativeElement, 'groupExpand', value);
     }
@@ -38,8 +34,6 @@ export class NimbleLabelProviderTableDirective {
         return this.elementRef.nativeElement.groupsCollapseAll;
     }
 
-    // Renaming because property should have camel casing, but attribute should not
-    // eslint-disable-next-line @angular-eslint/no-input-rename
     @Input('groups-collapse-all') public set groupsCollapseAll(value: string | undefined) {
         this.renderer.setProperty(this.elementRef.nativeElement, 'groupsCollapseAll', value);
     }
@@ -48,8 +42,6 @@ export class NimbleLabelProviderTableDirective {
         return this.elementRef.nativeElement.cellActionMenu;
     }
 
-    // Renaming because property should have camel casing, but attribute should not
-    // eslint-disable-next-line @angular-eslint/no-input-rename
     @Input('cell-action-menu') public set cellActionMenu(value: string | undefined) {
         this.renderer.setProperty(this.elementRef.nativeElement, 'cellActionMenu', value);
     }
@@ -58,8 +50,6 @@ export class NimbleLabelProviderTableDirective {
         return this.elementRef.nativeElement.columnHeaderGrouped;
     }
 
-    // Renaming because property should have camel casing, but attribute should not
-    // eslint-disable-next-line @angular-eslint/no-input-rename
     @Input('column-header-grouped') public set columnHeaderGrouped(value: string | undefined) {
         this.renderer.setProperty(this.elementRef.nativeElement, 'columnHeaderGrouped', value);
     }
@@ -68,8 +58,6 @@ export class NimbleLabelProviderTableDirective {
         return this.elementRef.nativeElement.columnHeaderSortedAscending;
     }
 
-    // Renaming because property should have camel casing, but attribute should not
-    // eslint-disable-next-line @angular-eslint/no-input-rename
     @Input('column-header-sorted-ascending') public set columnHeaderSortedAscending(value: string | undefined) {
         this.renderer.setProperty(this.elementRef.nativeElement, 'columnHeaderSortedAscending', value);
     }
@@ -78,8 +66,6 @@ export class NimbleLabelProviderTableDirective {
         return this.elementRef.nativeElement.columnHeaderSortedDescending;
     }
 
-    // Renaming because property should have camel casing, but attribute should not
-    // eslint-disable-next-line @angular-eslint/no-input-rename
     @Input('column-header-sorted-descending') public set columnHeaderSortedDescending(value: string | undefined) {
         this.renderer.setProperty(this.elementRef.nativeElement, 'columnHeaderSortedDescending', value);
     }
@@ -88,8 +74,6 @@ export class NimbleLabelProviderTableDirective {
         return this.elementRef.nativeElement.selectAll;
     }
 
-    // Renaming because property should have camel casing, but attribute should not
-    // eslint-disable-next-line @angular-eslint/no-input-rename
     @Input('select-all') public set selectAll(value: string | undefined) {
         this.renderer.setProperty(this.elementRef.nativeElement, 'selectAll', value);
     }
@@ -98,8 +82,6 @@ export class NimbleLabelProviderTableDirective {
         return this.elementRef.nativeElement.groupSelectAll;
     }
 
-    // Renaming because property should have camel casing, but attribute should not
-    // eslint-disable-next-line @angular-eslint/no-input-rename
     @Input('group-select-all') public set groupSelectAll(value: string | undefined) {
         this.renderer.setProperty(this.elementRef.nativeElement, 'groupSelectAll', value);
     }
@@ -108,8 +90,6 @@ export class NimbleLabelProviderTableDirective {
         return this.elementRef.nativeElement.rowSelect;
     }
 
-    // Renaming because property should have camel casing, but attribute should not
-    // eslint-disable-next-line @angular-eslint/no-input-rename
     @Input('row-select') public set rowSelect(value: string | undefined) {
         this.renderer.setProperty(this.elementRef.nativeElement, 'rowSelect', value);
     }
@@ -118,8 +98,6 @@ export class NimbleLabelProviderTableDirective {
         return this.elementRef.nativeElement.rowOperationColumn;
     }
 
-    // Renaming because property should have camel casing, but attribute should not
-    // eslint-disable-next-line @angular-eslint/no-input-rename
     @Input('row-operation-column') public set rowOperationColumn(value: string | undefined) {
         this.renderer.setProperty(this.elementRef.nativeElement, 'rowOperationColumn', value);
     }