+++ packages/nimble-components/src/rich-text-mention/users/view/tests/rich-text-mention-users-view.spec.ts
@@ -11,7 +11,6 @@ async function setup(): Promise<Fixture<RichTextMentionUsersView>> {
         html`<${richTextMentionUsersViewTag}
             mention-href="users:1"
             mention-label="John Doe"
-            disable-editing
             >@John Doe</${richTextMentionUsersViewTag}
         >`
     );
@@ -54,13 +53,13 @@ describe('RichTextMentionUsersView', () => {
     });
 
     describe('various wacky strings should reflect the `mention-label` attribute value to its text content', () => {
-        parameterizeNamedList(wackyStrings, (spec, name, value) => {
+        parameterizeNamedList(wackyStrings, (spec, name) => {
             spec(`for ${name}`, async () => {
                 await connect();
-                element.setAttribute('mention-label', value.name);
+                element.setAttribute('mention-label', name);
 
                 await waitForUpdatesAsync();
-                expect(pageObject.getTextContent()).toBe(`@${value.name}`);
+                expect(pageObject.getTextContent()).toBe(`@${name}`);
             });
         });
     });