+++ packages/nimble-components/src/table/index.ts
@@ -307,7 +307,9 @@ export class Table<
         return selectedRecordIds;
     }
 
-    public async setSelectedRecordIds(recordIds: string[]): Promise<void> {
+    public async setSelectedRecordIds(
+        recordIds: readonly string[]
+    ): Promise<void> {
         await this.processPendingUpdates();
 
         if (this.selectionMode === TableRowSelectionMode.none) {
@@ -324,7 +326,7 @@ export class Table<
     public override connectedCallback(): void {
         super.connectedCallback();
         this.initialize();
-        this.virtualizer.connectedCallback();
+        this.virtualizer.connect();
         this.viewport.addEventListener('scroll', this.onViewPortScroll, {
             passive: true
         });
@@ -334,7 +336,7 @@ export class Table<
 
     public override disconnectedCallback(): void {
         super.disconnectedCallback();
-        this.virtualizer.disconnectedCallback();
+        this.virtualizer.disconnect();
         this.viewport.removeEventListener('scroll', this.onViewPortScroll);
         document.removeEventListener('keydown', this.onKeyDown);
         document.removeEventListener('keyup', this.onKeyUp);
@@ -1075,7 +1077,7 @@ export class Table<
     }
 
     private calculateTanStackSelectionState(
-        recordIdsToSelect: string[]
+        recordIdsToSelect: readonly string[]
     ): TanStackRowSelectionState {
         if (this.selectionMode === TableRowSelectionMode.none) {
             return {};
@@ -1103,4 +1105,4 @@ const nimbleTable = Table.compose({
 });
 
 DesignSystem.getOrCreate().withPrefix('nimble').register(nimbleTable());
-export const tableTag = DesignSystem.tagFor(Table);
+export const tableTag = 'nimble-table';