+++ packages/nimble-components/src/text-field/index.ts
@@ -69,4 +69,4 @@ const nimbleTextField = TextField.compose<TextFieldOptions>({
 });
 
 DesignSystem.getOrCreate().withPrefix('nimble').register(nimbleTextField());
-export const textFieldTag = DesignSystem.tagFor(TextField);
+export const textFieldTag = 'nimble-text-field';